SKALE-4258 Increase coverage for depositboxERC721
