+++ proxy/test/utils/deploy/mainnet/linker.ts
@@ -8,5 +8,6 @@ export async function deployLinker(
     const factory = await ethers.getContractFactory("Linker");
     const instance = await factory.deploy() as Linker;
     await instance["initialize(address,address)"](contractManager.address, messageProxy.address);
+    await instance.registerMainnetContract(instance.address);
     return instance;
 }
